+++ src/js/components/dataDives/EquityCovidSpendingPage.jsx
@@ -12,8 +12,8 @@ import { ShareIcon } from 'data-transparency-ui';
 import { Link } from "react-router-dom";
 import PageWrapper from "../sharedComponents/PageWrapper";
 import { equityPageMetaTags } from "../../helpers/metaTagHelper";
-import MainCards from "./sections/MainCards";
-import EquityHeading from "./equity/shared/EquityHeading";
+import MainCards from "./shared/MainCards";
+import EquityHeading from "./shared/DataDiveHeading.jsx";
 import EquitySpotlightCards from "./equity/EquitySpotlightCards";
 
 require('pages/equityCovidSpendingPage/equityCovidSpendingPage.scss');
