+++ src/js/components/sharedComponents/cards/SpotlightCard.jsx
